@@ -1,29 +1,11 @@
-// Copyright (c) PlanB. GmbH. All Rights Reserved.
-// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
-
 using System;
 
 namespace PlanB.Butler.Admin.Models
 {
-    /// <summary>
-    /// ErrorViewModel.
-    /// </summary>
     public class ErrorViewModel
     {
-        /// <summary>
-        /// Gets or sets the request identifier.
-        /// </summary>
-        /// <value>
-        /// The request identifier.
-        /// </value>
         public string RequestId { get; set; }
 
-        /// <summary>
-        /// Gets a value indicating whether [show request identifier].
-        /// </summary>
-        /// <value>
-        ///   <c>true</c> if [show request identifier]; otherwise, <c>false</c>.
-        /// </value>
-        public bool ShowRequestId => !string.IsNullOrEmpty(this.RequestId);
+        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
     }
 }
