@@ -2,230 +2,7 @@
     ViewData["Title"] = "Home Page";
 }
 
-@if (ViewData["Message"] != null)
-{
-    if (ViewData["Message"].ToString() != "fail")
-    {
-        <div id="myModal" class="modal">
-            <!-- Modal content -->
-            <div class="modal-content">
-                <i class="fa fa-check-circle check"></i>
-            </div>
-        </div>
-    }
-    else
-    {
-        <div id="myModal" class="modal">
-            <!-- Modal content -->
-            <div class="modal-content text-center">
-                <i class="fa fa-times-circle check m-auto"></i>
-                <span>Either nothing is selected or an Exception was triggered</span>
-            </div>
-        </div>
-    }
-}
-<div class="container mt-3">
-    <form method="post">
-        <div class="w-100 container-row">
-            <span class="r-header">Bieg:</span>
-            <div class="float-right input-field w-50 hover">
-                <input class="form-control file-upload d-inline-block" placeholder="Import File..." id="file-selected-0" readonly />
-                <label for="file-upload-0" class="btn btn-orange">
-                    <i class="fa fa-plus-circle"></i>
-                    <input id="file-upload-0" type="file" name="filePath[]" accept="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet" />
-                </label>
-                <input class="form-control pic-upload d-inline-block" placeholder="Import File..." id="pic-selected-0" readonly />
-                <label for="pic-upload-0" class="btn btn-orange">
-                    <i class="fa fa-image"></i>
-                    <input id="pic-upload-0" type="file" name="picPath[]" accept="image/*" />
-                </label>
-            </div>
-        </div>
-        <div class="w-100 container-row">
-            <span class="r-header">Liederhalle:</span>
-            <div class="float-right input-field w-50 hover">
-                <input class="form-control file-upload d-inline-block" placeholder="Import File..." id="file-selected-1" readonly />
-                <label for="file-upload-1" class="btn btn-orange">
-                    <i class="fa fa-plus-circle"></i>
-                    <input id="file-upload-1" type="file" name="filePath[]" accept="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet" />
-                </label>
-                <input class="form-control pic-upload d-inline-block" placeholder="Import File..." id="pic-selected-1" readonly />
-                <label for="pic-upload-1" class="btn btn-orange">
-                    <i class="fa fa-image"></i>
-                    <input id="pic-upload-1" type="file" name="picPath[]" accept="image/*" />
-                </label>
-            </div>
-        </div>
-        <div class="w-100 container-row">
-            <span class="r-header">La Boussola:</span>
-            <div class="float-right input-field w-50 hover">
-                <input class="form-control file-upload d-inline-block" placeholder="Import File..." id="file-selected-2" readonly />
-                <label for="file-upload-2" class="btn btn-orange">
-                    <i class="fa fa-plus-circle"></i>
-                    <input id="file-upload-2" type="file" name="filePath[]" accept="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet" />
-                </label>
-                <input class="form-control pic-upload d-inline-block" placeholder="Import File..." id="pic-selected-2" readonly />
-                <label for="pic-upload-2" class="btn btn-orange">
-                    <i class="fa fa-image"></i>
-                    <input id="pic-upload-2" type="file" name="picPath[]" accept="image/*" />
-                </label>
-            </div>
-        </div>
-        <div class="w-100 container-row">
-            <span class="r-header">Leib und Seele:</span>
-            <div class="float-right input-field w-50 hover">
-                <input class="form-control file-upload d-inline-block" placeholder="Import File..." id="file-selected-3" readonly />
-                <label for="file-upload-3" class="btn btn-orange">
-                    <i class="fa fa-plus-circle"></i>
-                    <input id="file-upload-3" type="file" name="filePath[]" accept="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet" />
-                </label>
-                <input class="form-control pic-upload d-inline-block" placeholder="Import File..." id="pic-selected-3" readonly />
-                <label for="pic-upload-3" class="btn btn-orange">
-                    <i class="fa fa-image"></i>
-                    <input id="pic-upload-3" type="file" name="picPath[]" accept="image/*" />
-                </label>
-            </div>
-        </div>
-        <div class="w-100 container-row">
-            <span class="r-header">Delphi:</span>
-            <div class="float-right input-field w-50 hover">
-                <input class="form-control file-upload d-inline-block" placeholder="Import File..." id="file-selected-4" readonly />
-                <label for="file-upload-4" class="btn btn-orange">
-                    <i class="fa fa-plus-circle"></i>
-                    <input id="file-upload-4" type="file" name="filePath[]" accept="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet" />
-                </label>
-                <input class="form-control pic-upload d-inline-block" placeholder="Import File..." id="pic-selected-4" readonly />
-                <label for="pic-upload-4" class="btn btn-orange">
-                    <i class="fa fa-image"></i>
-                    <input id="pic-upload-4" type="file" name="picPath[]" accept="image/*" />
-                </label>
-            </div>
-        </div>
-        <div class="w-100 container-row">
-            <span class="r-header">Wilhelmshöhe:</span>
-            <div class="float-right input-field w-50 hover">
-                <input class="form-control file-upload d-inline-block" placeholder="Import File..." id="file-selected-5" readonly />
-                <label for="file-upload-5" class="btn btn-orange">
-                    <i class="fa fa-plus-circle"></i>
-                    <input id="file-upload-5" type="file" name="filePath[]" accept="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet" />
-                </label>
-                <input class="form-control pic-upload d-inline-block" placeholder="Import File..." id="pic-selected-5" readonly />
-                <label for="pic-upload-5" class="btn btn-orange">
-                    <i class="fa fa-image"></i>
-                    <input id="pic-upload-5" type="file" name="picPath[]" accept="image/*" />
-                </label>
-            </div>
-        </div>
-        <div class="w-100 container-row">
-            <span class="r-header">Feasy:</span>
-            <div class="float-right input-field w-50 hover">
-                <input class="form-control file-upload d-inline-block" placeholder="Import File..." id="file-selected-6" readonly />
-                <label for="file-upload-6" class="btn btn-orange">
-                    <i class="fa fa-plus-circle"></i>
-                    <input id="file-upload-6" type="file" name="filePath[]" accept="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet" />
-                </label>
-                <input class="form-control pic-upload d-inline-block" placeholder="Import File..." id="pic-selected-6" readonly />
-                <label for="pic-upload-6" class="btn btn-orange">
-                    <i class="fa fa-image"></i>
-                    <input id="pic-upload-6" type="file" name="picPath[]" accept="image/*" />
-                </label>
-            </div>
-        </div>
-        <input type="submit" id="submit" class="btn btn-orange w-100" href="/Home/Plan" />
-    </form>
+<div class="text-center">
+    <h1 class="display-4">Welcome</h1>
+    <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
 </div>
-<script>
-    $('#file-upload-0').bind('change', function () {
-
-        var fileName = ''; if ($(this).val() == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#file-selected-0').val(fileName);
-    })
-    $('#file-upload-1').bind('change', function () {
-        var fileName = ''; if ($(this).val == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#file-selected-1').val(fileName);
-    })
-    $('#file-upload-2').bind('change', function () {
-        var fileName = ''; if ($(this).val() == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#file-selected-2').val(fileName);
-    })
-    $('#file-upload-3').bind('change', function () {
-        var fileName = ''; if ($(this).val() == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#file-selected-3').val(fileName);
-    })
-    $('#file-upload-4').bind('change', function () {
-        var fileName = ''; if ($(this).val() == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#file-selected-4').val(fileName);
-    })
-    $('#file-upload-5').bind('change', function () {
-        var fileName = ''; if ($(this).val() == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#file-selected-5').val(fileName);
-    })
-    $('#file-upload-6').bind('change', function () {
-        var fileName = ''; if ($(this).val() == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#file-selected-6').val(fileName);
-    })
-
-    $('#pic-upload-0').bind('change', function () {
-
-        var fileName = ''; if ($(this).val() == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#pic-selected-0').val(fileName);
-    })
-    $('#pic-upload-1').bind('change', function () {
-        var fileName = ''; if ($(this).val == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#pic-selected-1').val(fileName);
-    })
-    $('#pic-upload-2').bind('change', function () {
-        var fileName = ''; if ($(this).val() == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#pic-selected-2').val(fileName);
-    })
-    $('#pic-upload-3').bind('change', function () {
-        var fileName = ''; if ($(this).val() == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#pic-selected-3').val(fileName);
-    })
-    $('#pic-upload-4').bind('change', function () {
-        var fileName = ''; if ($(this).val() == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#pic-selected-4').val(fileName);
-    })
-    $('#pic-upload-5').bind('change', function () {
-        var fileName = ''; if ($(this).val() == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#pic-selected-5').val(fileName);
-    })
-    $('#pic-upload-6').bind('change', function () {
-        var fileName = ''; if ($(this).val() == '') { fileName = 'No File selected' } else {
-            fileName = $(this).val();
-        }
-        $('#pic-selected-6').val(fileName);
-    })
-
-    $('#myModal').bind('click', function (event) {
-        if (event.target != $('.modal-content')) {
-            $('#myModal').delay(100).fadeOut(400)
-        }
-    })
-</script>
